perf: api calls should end with slash
The resource is {uri}/resource/, if we do {uri}/resource, FastAPI will
return a 307 redirect to {uri}/resource/.

So, this means another network call, which means app/web is slower.

Plus, in this case we didn't have the HTTPSRedirectMiddleware in
FastAPI, so the call redirected to http, which means we had CORS issues.

Hopefully now it is fixed, and remember, allways put slashes!